Fix flakiness in test_spans
